+++ startup/GafferScene/standardOptions.py
@@ -54,29 +54,29 @@ def __rendererPresetNames( additionalNames ) :
 
 Gaffer.Metadata.registerValues( {
 
-	"option:render:camera" : [
+	"option:render:camera" : {
 
-		"defaultValue", "",
-		"description",
+		"defaultValue" : "",
+		"description" :
 		"""
 		The primary camera to be used for rendering. If this
 		is not specified, then a default orthographic camera
 		positioned at the origin is used.
 		""",
-		"label", "Camera",
-		"layout:section", "Camera",
+		"label" : "Camera",
+		"layout:section" : "Camera",
 
-		"plugValueWidget:type", "GafferSceneUI.ScenePathPlugValueWidget",
-		"path:valid", True,
-		"scenePathPlugValueWidget:setNames", IECore.StringVectorData( [ "__cameras" ] ),
-		"scenePathPlugValueWidget:setsLabel", "Show only cameras",
+		"plugValueWidget:type" : "GafferSceneUI.ScenePathPlugValueWidget",
+		"path:valid" : True,
+		"scenePathPlugValueWidget:setNames" : IECore.StringVectorData( [ "__cameras" ] ),
+		"scenePathPlugValueWidget:setsLabel" : "Show only cameras",
 
-	],
+	},
 
-	"option:render:filmFit" : [
+	"option:render:filmFit" : {
 
-		"defaultValue", 0,
-		"description",
+		"defaultValue" : 0,
+		"description" :
 		"""
 		How the aperture gate (the frame defined by the aperture) will
 		fit into the resolution gate (the framed defined by the data
@@ -113,58 +113,58 @@ Gaffer.Metadata.registerValues( {
 		image will appear vertically or horizontally stretched or
 		squeezed.
 		""",
-		"label", "Film Fit",
-		"layout:section", "Camera",
+		"label" : "Film Fit",
+		"layout:section" : "Camera",
 
-		"plugValueWidget:type", "GafferUI.PresetsPlugValueWidget",
-		"presetNames", IECore.StringVectorData( [ "Horizontal", "Vertical", "Fit", "Fill", "Distort" ] ),
-		"presetValues", IECore.IntVectorData( [ IECoreScene.Camera.FilmFit.Horizontal, IECoreScene.Camera.FilmFit.Vertical, IECoreScene.Camera.FilmFit.Fit, IECoreScene.Camera.FilmFit.Fill, IECoreScene.Camera.FilmFit.Distort ] ),
+		"plugValueWidget:type" : "GafferUI.PresetsPlugValueWidget",
+		"presetNames" : IECore.StringVectorData( [ "Horizontal", "Vertical", "Fit", "Fill", "Distort" ] ),
+		"presetValues" : IECore.IntVectorData( [ IECoreScene.Camera.FilmFit.Horizontal, IECoreScene.Camera.FilmFit.Vertical, IECoreScene.Camera.FilmFit.Fit, IECoreScene.Camera.FilmFit.Fill, IECoreScene.Camera.FilmFit.Distort ] ),
 
-	],
+	},
 
-	"option:render:resolution" : [
+	"option:render:resolution" : {
 
-		"defaultValue", imath.V2i( 1024, 778 ),
-		"description",
+		"defaultValue" : imath.V2i( 1024, 778 ),
+		"description" :
 		"""
 		The resolution of the image to be rendered.
 		""",
-		"label", "Resolution",
-		"layout:section", "Camera",
+		"label" : "Resolution",
+		"layout:section" : "Camera",
 
-	],
+	},
 
-	"option:render:pixelAspectRatio" : [
+	"option:render:pixelAspectRatio" : {
 
-		"defaultValue", 1.0,
-		"description",
+		"defaultValue" : 1.0,
+		"description" :
 		"""
 		The `width / height` aspect ratio of the individual pixels in
 		the rendered image.
 		""",
-		"label", "Pixel Aspect Ratio",
-		"layout:section", "Camera",
+		"label" : "Pixel Aspect Ratio",
+		"layout:section" : "Camera",
 
-	],
+	},
 
-	"option:render:resolutionMultiplier" : [
+	"option:render:resolutionMultiplier" : {
 
-		"defaultValue", 1.0,
-		"description",
+		"defaultValue" : 1.0,
+		"description" :
 		"""
 		Multiplies the resolution of the render by this amount.
 		""",
-		"label", "Resolution Multiplier",
-		"layout:section", "Camera",
+		"label" : "Resolution Multiplier",
+		"layout:section" : "Camera",
 
-	],
+	},
 
-	"option:render:cropWindow" : [
+	"option:render:cropWindow" : {
 
-		"defaultValue", imath.Box2f( imath.V2f( 0, 0 ), imath.V2f( 1, 1 ) ),
-		"minValue", imath.V2f( 0, 0 ),
-		"maxValue", imath.V2f( 1, 1 ),
-		"description",
+		"defaultValue" : imath.Box2f( imath.V2f( 0, 0 ), imath.V2f( 1, 1 ) ),
+		"minValue" : imath.V2f( 0, 0 ),
+		"maxValue" : imath.V2f( 1, 1 ),
+		"description" :
 		"""
 		Limits the render to a region of the image. The rendered image
 		will have the same resolution as usual, but areas outside the
@@ -173,15 +173,15 @@ Gaffer.Metadata.registerValues( {
 		crop window tool in the viewer may be used to set this
 		interactively.
 		""",
-		"label", "Crop Window",
-		"layout:section", "Camera",
+		"label" : "Crop Window",
+		"layout:section" : "Camera",
 
-	],
+	},
 
-	"option:render:overscan" : [
+	"option:render:overscan" : {
 
-		"defaultValue", False,
-		"description",
+		"defaultValue" : False,
+		"description" :
 		"""
 		Whether to enable overscan, which adds extra pixels to the
 		sides of the rendered image.
@@ -192,106 +192,106 @@ Gaffer.Metadata.registerValues( {
 		`render:overscanRight` options to specify the amount of overscan on
 		each side of the image.
 		""",
-		"label", "Overscan",
-		"layout:section", "Camera",
+		"label" : "Overscan",
+		"layout:section" : "Camera",
 
-	],
+	},
 
-	"option:render:overscanTop" : [
+	"option:render:overscanTop" : {
 
-		"defaultValue", 0.0,
-		"minValue", 0.0,
-		"maxValue", 1.0,
-		"description",
+		"defaultValue" : 0.0,
+		"minValue" : 0.0,
+		"maxValue" : 1.0,
+		"description" :
 		"""
 		The amount of overscan at the top of the image. Specified as a
 		0-1 proportion of the original image height.
 		""",
-		"label", "Overscan Top",
-		"layout:section", "Camera",
+		"label" : "Overscan Top",
+		"layout:section" : "Camera",
 
-	],
+	},
 
-	"option:render:overscanBottom" : [
+	"option:render:overscanBottom" : {
 
-		"defaultValue", 0.0,
-		"minValue", 0.0,
-		"maxValue", 1.0,
-		"description",
+		"defaultValue" : 0.0,
+		"minValue" : 0.0,
+		"maxValue" : 1.0,
+		"description" :
 		"""
 		The amount of overscan at the bottom of the image. Specified as
 		a 0-1 proportion of the original image height.
 		""",
-		"label", "Overscan Bottom",
-		"layout:section", "Camera",
+		"label" : "Overscan Bottom",
+		"layout:section" : "Camera",
 
-	],
+	},
 
-	"option:render:overscanLeft" : [
+	"option:render:overscanLeft" : {
 
-		"defaultValue", 0.0,
-		"minValue", 0.0,
-		"maxValue", 1.0,
-		"description",
+		"defaultValue" : 0.0,
+		"minValue" : 0.0,
+		"maxValue" : 1.0,
+		"description" :
 		"""
 		The amount of overscan at the left of the image. Specified as a
 		0-1 proportion of the original image width.
 		""",
-		"label", "Overscan Left",
-		"layout:section", "Camera",
+		"label" : "Overscan Left",
+		"layout:section" : "Camera",
 
-	],
+	},
 
-	"option:render:overscanRight" : [
+	"option:render:overscanRight" : {
 
-		"defaultValue", 0.0,
-		"minValue", 0.0,
-		"maxValue", 1.0,
-		"description",
+		"defaultValue" : 0.0,
+		"minValue" : 0.0,
+		"maxValue" : 1.0,
+		"description" :
 		"""
 		The amount of overscan at the right of the image. Specified as
 		a 0-1 proportion of the original image width.
 		""",
-		"label", "Overscan Right",
-		"layout:section", "Camera",
+		"label" : "Overscan Right",
+		"layout:section" : "Camera",
 
-	],
+	},
 
-	"option:render:depthOfField" : [
+	"option:render:depthOfField" : {
 
-		"defaultValue", False,
-		"description",
+		"defaultValue" : False,
+		"description" :
 		"""
 		Whether to render with depth of field. To ensure the effect is
 		visible, you must also set an f-stop value greater than 0 on
 		this camera.
 		""",
-		"label", "Depth Of Field",
-		"layout:section", "Camera",
+		"label" : "Depth Of Field",
+		"layout:section" : "Camera",
 
-	],
+	},
 
-	"option:render:defaultRenderer" : [
+	"option:render:defaultRenderer" : {
 
-		"defaultValue", "",
-		"description",
+		"defaultValue" : "",
+		"description" :
 		"""
 		Specifies the default renderer to be used by the Render and
 		InteractiveRender nodes.
 		""",
-		"label", "Default Renderer",
-		"layout:section", "Renderer",
+		"label" : "Default Renderer",
+		"layout:section" : "Renderer",
 
-		"plugValueWidget:type", "GafferUI.PresetsPlugValueWidget",
-		"presetNames", __rendererPresetNames( [ "None" ] ),
-		"presetValues", __rendererPresetNames( [ "" ] ),
+		"plugValueWidget:type" : "GafferUI.PresetsPlugValueWidget",
+		"presetNames" : __rendererPresetNames( [ "None" ] ),
+		"presetValues" : __rendererPresetNames( [ "" ] ),
 
-	],
+	},
 
-	"option:render:includedPurposes" : [
+	"option:render:includedPurposes" : {
 
-		"defaultValue", IECore.StringVectorData( [ "default", "render" ] ),
-		"description",
+		"defaultValue" : IECore.StringVectorData( [ "default", "render" ] ),
+		"description" :
 		"""
 		Limits the objects included in the render according to the values of their `usd:purpose`
 		attribute. The "Default" purpose includes all objects which have no `usd:purpose` attribute;
@@ -299,17 +299,17 @@ Gaffer.Metadata.registerValues( {
 
 		> Tip : Use the USDAttributes node to assign the `usd:purpose` attribute.
 		""",
-		"label", "Included Purposes",
-		"layout:section", "Render Set",
+		"label" : "Included Purposes",
+		"layout:section" : "Render Set",
 
-		"plugValueWidget:type", "GafferSceneUI.StandardOptionsUI._IncludedPurposesPlugValueWidget",
+		"plugValueWidget:type" : "GafferSceneUI.StandardOptionsUI._IncludedPurposesPlugValueWidget",
 
-	],
+	},
 
-	"option:render:inclusions" : [
+	"option:render:inclusions" : {
 
-		"defaultValue", "/",
-		"description",
+		"defaultValue" : "/",
+		"description" :
 		"""
 		A set expression that limits the objects included in the render to only those matched
 		and their descendants. Objects not matched by the set expression will be pruned from
@@ -317,52 +317,52 @@ Gaffer.Metadata.registerValues( {
 
 		> Tip : Cameras are included by default and do not need to be specified here.
 		""",
-		"label", "Inclusions",
-		"layout:section", "Render Set",
+		"label" : "Inclusions",
+		"layout:section" : "Render Set",
 
-		"plugValueWidget:type", "GafferSceneUI.SetExpressionPlugValueWidget",
-		"ui:scene:acceptsSetExpression", True,
+		"plugValueWidget:type" : "GafferSceneUI.SetExpressionPlugValueWidget",
+		"ui:scene:acceptsSetExpression" : True,
 
-	],
+	},
 
-	"option:render:exclusions" : [
+	"option:render:exclusions" : {
 
-		"defaultValue", "",
-		"description",
+		"defaultValue" : "",
+		"description" :
 		"""
 		A set expression that excludes the matched objects from the render. Exclusions
 		affect both `inclusions` and `additionalLights` and cause the matching objects and
 		their descendants to be pruned from the scene.
 		""",
-		"label", "Exclusions",
-		"layout:section", "Render Set",
+		"label" : "Exclusions",
+		"layout:section" : "Render Set",
 
-		"plugValueWidget:type", "GafferSceneUI.SetExpressionPlugValueWidget",
-		"ui:scene:acceptsSetExpression", True,
+		"plugValueWidget:type" : "GafferSceneUI.SetExpressionPlugValueWidget",
+		"ui:scene:acceptsSetExpression" : True,
 
-	],
+	},
 
-	"option:render:additionalLights" : [
+	"option:render:additionalLights" : {
 
-		"defaultValue", "",
-		"description",
+		"defaultValue" : "",
+		"description" :
 		"""
 		A set expression that specifies additional lights to be included in the render.
 		This differs from `inclusions` in that only lights and light filters will be
 		matched by this set expression.
 		""",
-		"label", "Additional Lights",
-		"layout:section", "Render Set",
+		"label" : "Additional Lights",
+		"layout:section" : "Render Set",
 
-		"plugValueWidget:type", "GafferSceneUI.SetExpressionPlugValueWidget",
-		"ui:scene:acceptsSetExpression", True,
+		"plugValueWidget:type" : "GafferSceneUI.SetExpressionPlugValueWidget",
+		"ui:scene:acceptsSetExpression" : True,
 
-	],
+	},
 
-	"option:render:cameraInclusions" : [
+	"option:render:cameraInclusions" : {
 
-		"defaultValue", "/",
-		"description",
+		"defaultValue" : "/",
+		"description" :
 		"""
 		A set expression that limits the objects visible to camera rays to only those matched
 		and their descendants. Camera visibility attributes authored in the scene take
@@ -376,18 +376,18 @@ Gaffer.Metadata.registerValues( {
 		# OS & desktop environment, such that some users see the whole label, while it is truncated
 		# for others. We should look into whether shipping a standard font with Gaffer would reduce
 		# this ambiguity.
-		"label", "Camera Inclusions",
-		"layout:section", "Visibility Set",
+		"label" : "Camera Inclusions",
+		"layout:section" : "Visibility Set",
 
-		"plugValueWidget:type", "GafferSceneUI.SetExpressionPlugValueWidget",
-		"ui:scene:acceptsSetExpression", True,
+		"plugValueWidget:type" : "GafferSceneUI.SetExpressionPlugValueWidget",
+		"ui:scene:acceptsSetExpression" : True,
 
-	],
+	},
 
-	"option:render:cameraExclusions" : [
+	"option:render:cameraExclusions" : {
 
-		"defaultValue", "",
-		"description",
+		"defaultValue" : "",
+		"description" :
 		"""
 		A set expression that excludes the matched objects and their descendants from camera
 		ray visibility. Camera visibility attributes authored in the scene take precedence
@@ -401,35 +401,35 @@ Gaffer.Metadata.registerValues( {
 		cast shadows or reflections. Shadow or reflection visibility attributes authored
 		in the scene take precedence over this option.
 		""",
-		"label", "Camera Exclusions",
-		"layout:section", "Visibility Set",
+		"label" : "Camera Exclusions",
+		"layout:section" : "Visibility Set",
 
-		"plugValueWidget:type", "GafferSceneUI.SetExpressionPlugValueWidget",
-		"ui:scene:acceptsSetExpression", True,
+		"plugValueWidget:type" : "GafferSceneUI.SetExpressionPlugValueWidget",
+		"ui:scene:acceptsSetExpression" : True,
 
-	],
+	},
 
-	"option:render:matteInclusions" : [
+	"option:render:matteInclusions" : {
 
-		"defaultValue", "",
-		"description",
+		"defaultValue" : "",
+		"description" :
 		"""
 		A set expression that specifies objects that should be treated as matte (holdout)
 		objects along with their descendants. Matte attributes authored in the scene take
 		precedence over this option.
 		""",
-		"label", "Matte Inclusions",
-		"layout:section", "Visibility Set",
+		"label" : "Matte Inclusions",
+		"layout:section" : "Visibility Set",
 
-		"plugValueWidget:type", "GafferSceneUI.SetExpressionPlugValueWidget",
-		"ui:scene:acceptsSetExpression", True,
+		"plugValueWidget:type" : "GafferSceneUI.SetExpressionPlugValueWidget",
+		"ui:scene:acceptsSetExpression" : True,
 
-	],
+	},
 
-	"option:render:matteExclusions" : [
+	"option:render:matteExclusions" : {
 
-		"defaultValue", "",
-		"description",
+		"defaultValue" : "",
+		"description" :
 		"""
 		A set expression that excludes the matched objects and their descendants from being
 		treated as matte (holdout) objects. Matte attributes authored in the scene take
@@ -439,18 +439,18 @@ Gaffer.Metadata.registerValues( {
 		as locations not specified in `matteInclusions` already default to not being
 		treated as matte objects.
 		""",
-		"label", "Matte Exclusions",
-		"layout:section", "Visibility Set",
+		"label" : "Matte Exclusions",
+		"layout:section" : "Visibility Set",
 
-		"plugValueWidget:type", "GafferSceneUI.SetExpressionPlugValueWidget",
-		"ui:scene:acceptsSetExpression", True,
+		"plugValueWidget:type" : "GafferSceneUI.SetExpressionPlugValueWidget",
+		"ui:scene:acceptsSetExpression" : True,
 
-	],
+	},
 
-	"option:render:transformBlur" : [
+	"option:render:transformBlur" : {
 
-		"defaultValue", False,
-		"description",
+		"defaultValue" : False,
+		"description" :
 		"""
 		Whether or not transform motion is taken into
 		account in the rendered image. To specify the
@@ -458,15 +458,15 @@ Gaffer.Metadata.registerValues( {
 		object in the scene, use a StandardAttributes
 		node with appropriate filters.
 		""",
-		"label", "Transform",
-		"layout:section", "Motion Blur",
+		"label" : "Transform",
+		"layout:section" : "Motion Blur",
 
-	],
+	},
 
-	"option:render:deformationBlur" : [
+	"option:render:deformationBlur" : {
 
-		"defaultValue", False,
-		"description",
+		"defaultValue" : False,
+		"description" :
 		"""
 		Whether or not deformation motion is taken into
 		account in the rendered image. To specify the
@@ -474,70 +474,70 @@ Gaffer.Metadata.registerValues( {
 		object in the scene, use a StandardAttributes
 		node with appropriate filters.
 		""",
-		"label", "Deformation",
-		"layout:section", "Motion Blur",
+		"label" : "Deformation",
+		"layout:section" : "Motion Blur",
 
-	],
+	},
 
-	"option:render:shutter" : [
+	"option:render:shutter" : {
 
-		"defaultValue", imath.V2f( -0.25, 0.25 ),
-		"description",
+		"defaultValue" : imath.V2f( -0.25, 0.25 ),
+		"description" :
 		"""
 		The interval over which the camera shutter is open. Measured
 		in frames, and specified relative to the frame being rendered.
 		""",
-		"label", "Shutter",
-		"layout:section", "Motion Blur",
+		"label" : "Shutter",
+		"layout:section" : "Motion Blur",
 
-	],
+	},
 
-	"option:sampleMotion" : [
+	"option:sampleMotion" : {
 
-		"defaultValue", True,
-		"description",
+		"defaultValue" : True,
+		"description" :
 		"""
 		Whether to actually render motion blur. Disabling this
 		setting while motion blur is set up produces a render where
 		there is no blur, but there is accurate motion information.
 		Useful for rendering motion vector passes.
 		""",
-		"label", "Sample Motion",
-		"layout:section", "Motion Blur",
+		"label" : "Sample Motion",
+		"layout:section" : "Motion Blur",
 
-	],
+	},
 
-	"option:render:manifestFilePath" : [
+	"option:render:manifestFilePath" : {
 
-		"defaultValue", "",
-		"description",
+		"defaultValue" : "",
+		"description" :
 		"""
 		Specifies a file to write a matching ID manifest to, when
 		rendering an ID aov in a batch render. This is needed to use
 		the Image Selection Tool with batch renders (interactive
 		renders just need an ID aov).
 		""",
-		"label", "File Path",
-		"layout:section", "Render Manifest",
+		"label" : "File Path",
+		"layout:section" : "Render Manifest",
 
-		"plugValueWidget:type", "GafferUI.FileSystemPathPlugValueWidget",
-		"path:leaf", True,
-		"fileSystemPath:extensions", "exr",
+		"plugValueWidget:type" : "GafferUI.FileSystemPathPlugValueWidget",
+		"path:leaf" : True,
+		"fileSystemPath:extensions" : "exr",
 
-	],
+	},
 
-	"option:render:performanceMonitor" : [
+	"option:render:performanceMonitor" : {
 
-		"defaultValue", False,
-		"description",
+		"defaultValue" : False,
+		"description" :
 		"""
 		Enables a performance monitor and uses it to output
 		statistics about scene generation performance.
 		""",
-		"label", "Performance Monitor",
-		"layout:section", "Statistics",
+		"label" : "Performance Monitor",
+		"layout:section" : "Statistics",
 
-	],
+	},
 
 } )